Change: optimized codes

+++ Program.cs
@@ -3,6 +3,7 @@ using System.Globalization;
 using System.IO;
 using System.IO.Packaging;
 using System.Linq;
+using System.Net.WebSockets;
 using System.Text;
 
 using Mono.Options;
@@ -10,10 +11,13 @@ using OpenMcdf;
 
 public class Program
 {
+    public enum OfficeFileFormat { Unknown, ZIP, MCDF };
+
     private static string AppName = Path.GetFileName(AppDomain.CurrentDomain.FriendlyName);
 
     private static Encoding MBCS = Encoding.Default;
 
+    private static bool overwrite = false;
     private static bool extract_all = true;
     private static bool extract_media = false;
     private static bool extract_embed = false;
@@ -27,6 +31,7 @@ public class Program
     private static OptionSet Options { get; set; } = new OptionSet()
     {
         { "h|?|help", "Show Help", v => { ShowHelp(); } },
+        { "y|overwrite", "Overwrite exists file", v => { overwrite = v != null; } },
         { "a|all", "Extracting media and embeddings", v => { extract_all = v != null; } },
         { "m|media", "Extracting media ", v => { extract_media = v != null; } },
         { "e|object", "Extracting embeddings", v => { extract_embed = v != null; } },
@@ -52,7 +57,11 @@ public class Program
         var doc = opts.FirstOrDefault();
         if (string.IsNullOrEmpty(doc) || !File.Exists(doc)) { ShowHelp(); return; }
 
-        Export(doc);
+        try
+        {
+            Export(doc);
+        }
+        catch(Exception ex) { Console.WriteLine(ex.Message + Environment.NewLine + ex.StackTrace); };
     }
 
     public static void ShowHelp()
@@ -70,77 +79,97 @@ public class Program
         }
     }
 
-    private static void ExportOle10Native(byte[] data, string targetFolder)
+    private static bool ExportOle10Native(byte[] data, string targetFolder)
     {
-        if (data is byte[] && data.Length > 0 && !string.IsNullOrEmpty(targetFolder) && Directory.Exists(targetFolder))
+        var result = false;
+        Func<Stream, int, string> GetString = (stream, count) =>
+        {
+            var result = string.Empty;
+            if (stream is Stream && count > 0)
+            {
+                var pos = stream.Position;
+                var buf = new byte[count];
+                stream.Read(buf, 0, (int)Math.Max(0, Math.Min(stream.Length - pos, count)));
+                var bytes = buf.TakeWhile(b => b != 0).ToArray();
+                if (bytes.Length > 0)
+                {
+                    result = MBCS.GetString(bytes).Trim();
+                    stream.Seek(pos + bytes.Length, SeekOrigin.Begin);
+                    //Console.WriteLine($"{pos:08N} : {result}");
+                }
+            }
+            return(result);
+        };
+        Func<Stream, int> GetInt32 = stream =>
+        {
+            var result = 0;
+            if (stream is Stream)
+            {
+                var pos = stream.Position;
+                var buf = new byte[4];
+                stream.Read(buf, 0, (int)Math.Max(0, Math.Min(stream.Length - pos, 4)));
+                result = BitConverter.ToInt32(buf, 0);
+            }
+            return(result);
+        };
+
+        if (data?.Length > 0 && !string.IsNullOrEmpty(targetFolder) && Directory.Exists(targetFolder))
         {
             using (var olems = new MemoryStream(data))
             {
                 //Console.WriteLine(olems.Length);
-
                 var pos = olems.Position;
 
                 olems.Seek(pos + 6, SeekOrigin.Begin);
+                var filename = GetString(olems, 260);
                 pos = olems.Position;
-                var fn0 = string.Empty;
-                var f0 = new byte[260];
-                olems.Read(f0, 0, (int)Math.Min(olems.Length, 260));
-                var n0 = f0.TakeWhile(b => b != 0).ToArray();
-                //Console.WriteLine(n0.Length);
-                if (n0.Length > 0)
-                {
-                    fn0 = MBCS.GetString(n0).Trim();
-                    //Console.WriteLine(fn0);
-                }
-                pos += n0.Length;
 
                 olems.Seek(pos + 1, SeekOrigin.Begin);
+                var src_file = GetString(olems, 260);
                 pos = olems.Position;
-                var fn1 = string.Empty;
-                var f1 = new byte[260];
-                olems.Read(f1, 0, (int)Math.Min(olems.Length, 260));
-                var n1 = f1.TakeWhile(b => b != 0).ToArray();
-                //Console.WriteLine(n1.Length);
-                if (n1.Length > 0)
-                {
-                    fn1 = MBCS.GetString(n1).Trim();
-                    //Console.WriteLine(fn1);
-                }
-                pos += n1.Length + 4;
+
+                olems.Seek(pos + 4 + 1, SeekOrigin.Begin);
+                var temp_len = GetInt32(olems);
+                var temp_file = GetString(olems, temp_len);
+                pos = olems.Position;
 
                 olems.Seek(pos + 1, SeekOrigin.Begin);
-                pos = olems.Position;
-                var s2 = new byte[4];
-                olems.Read(s2, 0, 4);
-                var l2 = BitConverter.ToInt32(s2, 0);
-                //Console.WriteLine(l2);
-                var fn2 = string.Empty;
-                var f2 = new byte[l2];
-                olems.Read(f2, 0, l2);
-                var n2 = f2.TakeWhile(b => b != 0).ToArray();
-                //Console.WriteLine(n2.Length);
-                if (n2.Length > 0)
-                {
-                    fn2 = MBCS.GetString(n2).Trim();
-                    //Console.WriteLine(fn2);
-                }
-                pos += n2.Length;
-
-                pos = olems.Position;
-                //Console.WriteLine(pos);
-                var cs = new byte[4];
-                olems.Read(cs, 0, 4);
-                var cl = BitConverter.ToInt32(cs, 0);
-                //Console.WriteLine(cl);
+                var cl = GetInt32(olems);
                 var cnt = new byte[cl];
                 olems.Read(cnt, 0, cl);
 
-                var targetFile = Path.Combine(targetFolder, fn0);
-                Console.WriteLine("File : " + targetFile);
-                File.WriteAllBytes(targetFile, cnt);
-                Console.WriteLine("");
+                var targetFile = Path.Combine(targetFolder, filename);
+                Console.WriteLine("File  : " + targetFile);
+                if (overwrite || !File.Exists(targetFile))
+                {
+                    File.WriteAllBytes(targetFile, cnt);
+                    result = true;
+                }
             }
         }
+        return (result);
+    }
+
+    public static OfficeFileFormat DetectOfficeFileFormat(string officeFile)
+    {
+        var result = OfficeFileFormat.Unknown;
+        if (!string.IsNullOrEmpty(officeFile) && File.Exists(officeFile))
+        {
+            using (var fs = File.OpenRead(officeFile))
+            {
+                var buf = new byte[4096];
+                var ret = fs.Read(buf, 0, buf.Length);
+                //fs.ReadExactly(buf, 0, buf.Length);
+                if (ret > 0)
+                {
+                    if (buf[0] == 'P' && buf[1] == 'K' && buf[2] == 0x03 && buf[3] == 0x04)
+                        result = OfficeFileFormat.ZIP;
+                    else if (buf[0] == 0xD0 && buf[1] == 0xCF && buf[2] == 0x11 && buf[3] == 0xE0 && buf[4] == 0xA1 && buf[5] == 0xB1 && buf[6] == 0x1A && buf[7] == 0xE1)
+                        result = OfficeFileFormat.MCDF;
+                }
+            }
+        }
+        return (result);
     }
 
     public static void Export(string officeFile)
@@ -153,8 +182,9 @@ public class Program
         else if (officeExt.StartsWith(".xls")) office_type = "excel";
         else if (officeExt.StartsWith(".ppt")) office_type = "ppt";
 
-        if (officeExt.EndsWith("x") || officeExt.EndsWith("m")) ExportX(officeFile);
-        else ExportC(officeFile);
+        var file_type = DetectOfficeFileFormat(officeFile);
+        if (file_type == OfficeFileFormat.ZIP) ExportX(officeFile);
+        else if (file_type == OfficeFileFormat.MCDF) ExportC(officeFile);
     }
 
     public static void ExportC(string officeFile)
@@ -164,17 +194,10 @@ public class Program
             using (var cf = new CompoundFile(officeFile))
             {
                 var action = false;
-
-                //var data = cf.RootStorage.GetStream("Data");
-                //var buf = data.GetData();
-                //var objs = cf.RootStorage.GetStorage("ObjectPool");
-                //var items = cf.GetAllNamedEntries();
-
                 if (extract_all || extract_media)
                 {
                     Console.WriteLine("-------------------------------------------------------------------------------");
                     Console.WriteLine(Path.Combine(output_folder_info.FullName, "media"));
-                    Console.WriteLine("");
 
                     ExportMediasC(cf);
 
@@ -184,7 +207,6 @@ public class Program
                 {
                     Console.WriteLine("-------------------------------------------------------------------------------");
                     Console.WriteLine(Path.Combine(output_folder_info.FullName, "embeddings"));
-                    Console.WriteLine("");
 
                     ExportAttachmentsC(cf);
 
@@ -199,12 +221,13 @@ public class Program
     {
         if (officeFile is CompoundFile)
         {
-            var Data = officeFile.RootStorage.GetStream("Data");
-            var buf = Data.GetData();
             Console.WriteLine($"Sorry, unsupported extracting the media files");
             return;
 
 #if DEBUG
+            var Data = officeFile.RootStorage.GetStream("Data");
+            var buf = Data.GetData();
+
             for (int sid = 0; sid < officeFile.GetNumDirectories(); sid++)
             {
                 var name = officeFile.GetNameDirEntry(sid);
@@ -241,7 +264,9 @@ public class Program
                     var targetFolder_info = Directory.CreateDirectory(targetFolder);
                     if (targetFolder_info.Exists)
                     {
-                        ExportOle10Native(data, targetFolder);
+                        Console.WriteLine();
+                        var ret = ExportOle10Native(data, targetFolder);
+                        Console.WriteLine($"State : Save {(ret ? "successful" : "failed, file already exists")}");
                     }
                 }
             }
@@ -258,7 +283,6 @@ public class Program
             {
                 Console.WriteLine("-------------------------------------------------------------------------------");
                 Console.WriteLine(Path.Combine(output, "media"));
-                Console.WriteLine("");
 
                 ExportMediasX(officeFile);
 
@@ -268,7 +292,6 @@ public class Program
             {
                 Console.WriteLine("-------------------------------------------------------------------------------");
                 Console.WriteLine(Path.Combine(output, "embeddings"));
-                Console.WriteLine("");
 
                 ExportAttachmentsX(officeFile);
 
@@ -300,20 +323,23 @@ public class Program
             {
                 foreach (var mediaPart in officeFile.GetParts().Where(f => f.Uri.ToString().StartsWith($"/{office_type}/media/")))
                 {
-                    Console.WriteLine("Uri  : " + mediaPart.Uri);
-                    //Console.WriteLine(mediaPart.ContentType);
-                    //Console.WriteLine(mediaPart.Package.PackageProperties.Language);
-                    using (var ms = mediaPart.GetStream())
+                    var targetFile = Path.Combine(targetFolder, Path.GetFileName(mediaPart.Uri.ToString()));
+                    if (overwrite || !File.Exists(targetFile))
                     {
-                        var buffer = new byte[ms.Length];
-                        ms.Read(buffer, 0, (int)ms.Length);
-                        ms.Close();
+                        Console.WriteLine();
+                        Console.WriteLine("Uri  : " + mediaPart.Uri);
+                        using (var ms = mediaPart.GetStream())
+                        {
+                            var buffer = new byte[ms.Length];
+                            ms.Read(buffer, 0, (int)ms.Length);
+                            ms.Close();
 
-                        var targetFile = Path.Combine(targetFolder, Path.GetFileName(mediaPart.Uri.ToString()));
-                        Console.WriteLine("File : " + targetFile);
-                        File.WriteAllBytes(targetFile, buffer);
-                        Console.WriteLine("");
+                            Console.WriteLine("File  : " + targetFile);
+                            File.WriteAllBytes(targetFile, buffer);
+                            Console.WriteLine($"State : Save successful");
+                        }
                     }
+                    else Console.WriteLine($"State : Save failed, file already exists");
                 }
             }
         }
@@ -343,7 +369,6 @@ public class Program
                 foreach (var mediaPart in officeFile.GetParts().Where(f => f.Uri.ToString().StartsWith($"/{office_type}/embeddings/")))
                 {
                     //Console.WriteLine(mediaPart.ContentType);
-                    Console.WriteLine("Uri  : " + mediaPart.Uri);
                     using (var ms = mediaPart.GetStream())
                     {
                         ms.Seek(0, SeekOrigin.Begin);
@@ -367,8 +392,11 @@ public class Program
 
                                 if (name.EndsWith("Ole10Native"))
                                 {
+                                    Console.WriteLine();
+                                    Console.WriteLine("Uri   : " + mediaPart.Uri);
                                     var data = compoundFile.GetDataBySID(sid);
-                                    ExportOle10Native(data, targetFolder);
+                                    var ret = ExportOle10Native(data, targetFolder);
+                                    Console.WriteLine($"State : Save {(ret ? "successful" : "failed, file already exists")}");
                                 }
                             }
                         }
